Tower of Hanoi
Final Condition: Each disks is attached to the right most stick in a decsending order
from bottom to top.
Rules:
o Each move consists of taking the upper disk from one of the stacks and placing
it on top of another stack or an empty rod.
o No disk can be placed on top of the smaller disk.
o Only one disk may be moved at a time.
The program will ask for a number N (2 <= N <= 16), where N is the number of disks.
Then you have to start solving the game and print out the moves after each step with an
ordered tuple of lists. Each list represents the state of a single rod and the first item in
the list will be the first disks number from the bottom of the rod. For example, the
following state can be represented: